Fix rename worker message types.
