--- packages/engine/src/GameEngine/UserAIRuner/WorkerSupportTypes.ts
@@ -1,5 +1,0 @@
-//When adding constants update the function GetConstants() and getWorkerInitFunction()
-export const InitRequestType = 'InitRequest'
-export const InitAnswerType = 'InitAnswer'
-export const TurnRequestType = 'TurnRequest'
-export const TurnAnswerType = 'TurnAnswer'
